Refactor Tauri detection and fix shell plugin configuration
- Create shared Tauri detection utility in tauriUtils.ts with comprehensive detection logic
- Update all files to use shared Tauri detection instead of inconsistent inline checks
- Fix name collision in pdfExport.ts by removing redundant static isTauri() method
- Add shell plugin permissions to tauri.conf.json to enable external URL opening
- Improve license validation logic and remove early-return issue in licenseManager.ts
- Standardize Tauri detection across navigationUtils, licenseManager, pdfExport, and other components
