+++ src-tauri/tauri.conf.json
@@ -47,6 +47,9 @@
     ]
   },
   "plugins": {
+    "shell": {
+      "open": true
+    },
     "deep-link": {
       "desktop": {
         "schemes": ["leedpdf"]
